UPD: review fixes
