+++ src/db/UsersRepo.ts
@@ -1,8 +1,8 @@
 import { Pool } from "pg";
 import { IUsersRepo } from "../db/IUsersRepo";
 import { ConnParams } from "../types/ConnParams";
-import { validateUser } from "../validators";
-import { buildConn, q, USERS_TABLE } from "./common";
+import { buildConn, executeQuery, USERS_TABLE } from "./common";
+import { validateUser } from "./validateUser";
 import { User } from "../model/User";
 
 export class UsersRepo implements IUsersRepo {
@@ -14,7 +14,7 @@ export class UsersRepo implements IUsersRepo {
 
   async userExists(login: string) {
     const query = `SELECT * FROM ${USERS_TABLE} where login = '${login}';`;
-    const queryRes = await q(query, this.conn);
+    const queryRes = await executeQuery(query, this.conn);
 
     if (queryRes && queryRes.rows.length === 0) {
       return false;
@@ -39,7 +39,7 @@ export class UsersRepo implements IUsersRepo {
     const query = `INSERT INTO ${USERS_TABLE} (login, password, privelegelevel) VALUES \
     ('${user.login}', '${user.password}', '${user.privelegelevel}');`;
 
-    const qres = await q(query, this.conn);
+    const qres = await executeQuery(query, this.conn);
 
     return qres ? true : false;
   }
@@ -52,7 +52,7 @@ export class UsersRepo implements IUsersRepo {
     }
 
     const query = `SELECT * FROM ${USERS_TABLE} WHERE login = '${login}' AND password = '${password}';`;
-    const res = await q(query, this.conn);
+    const res = await executeQuery(query, this.conn);
     return res
       ? new User(res.rows[0].id, res.rows[0].login, "", res.rows[0].privelegelevel)
       : null;
