+++ src/types/Error.ts
@@ -1,0 +1,1 @@
+export type Error = { message: string }
