+++ src/types/Validation.ts
@@ -1,0 +1,5 @@
+import { Error } from "./Error"
+
+export type ValidationResult =
+  | { ok: true, errors: [] }
+  | { ok: false, errors: Error[] | [] }
